requirements.txt
streamlit>=1.34.0
openai>=1.30.0
chromadb>=0.5.3
python-dotenv>=1.0.1